% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circleclust.R
\name{circleclust}
\alias{circleclust}
\title{Cluster coordinates based on circular variance}
\usage{
circleclust(
  df,
  dt_field = NULL,
  circvar_threshold = 0.7,
  window = 60,
  cluster_threshold = NULL,
  show_circvar = FALSE
)
}
\arguments{
\item{df}{a data frame with columns \code{speed} and \code{azimuth} created by \code{move()}.
The data frame must also include datetime, longitude, and latitude colums.}

\item{circvar_threshold}{numeric; Circular variance threshold to determine clustering.
Default = 0.7.}

\item{window}{numeric; Window (number of rows) in which to calculate circular variance.}

\item{cluster_threshold}{numeric; The minimum allowable number of observations
in each cluster.  If the number of observations in an identified cluster is less than this threshold,
the observations are retained but not assigned a cluster value.}

\item{show_circvar}{logical; if \code{TRUE}, a column will be added to the output data
frame listing the circular variance (\code{circvar}) for each observation.
Default = \code{FALSE}.}
}
\value{
a tibble.
}
\description{
\code{circleclust} calculates the circular variance of consecutive latitude and
longitude coordinates within a moving window and identifies spatiotemporal clusters
based on a set of flexible parameters.  These parameters include the size of the
window in which to calculate circular variance, the threshold value (circular variance)
to identify clustering, and the minimum number of observations allowed in a
cluster.
}
\details{
Imputing lon/lat values using \code{impute_coords()} is recommended if GPS
coordinates are missing.
}
\examples{
\dontrun{

circleclust(df,
  circvar_threshold = .7, window = 60, cluster_threshold = NULL,
  show_circvar = FALSE
)
}
}
