% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_deck.R
\name{cluster_deck}
\alias{cluster_deck}
\title{Create a \code{mapdeck} pointcloud}
\usage{
cluster_deck(
  df,
  tooltip = "sp_temporal_cluster",
  fill_colour = "sp_temporal_cluster",
  elevation = NULL,
  radius = 10,
  palette = "viridis",
  legend = TRUE
)
}
\arguments{
\item{df}{an sf object}

\item{tooltip}{character; name of column or HTML text to render as a tooltip.}

\item{fill_colour}{character; name of column or hex color used to fill points.}

\item{elevation}{character; name of column with elevation values.}

\item{radius}{numeric; value in pixels of each point.}

\item{palette}{string or matrix; indicates a color palette.}

\item{legend}{logical; either a logical indicating if the legend should be
displayed, or a named list indicating which color attributes should be included
in the legend.}
}
\description{
'cluster_deck' is a convenience function to create a pointcloud map to
visualize clustered coordinates.
}
\details{
See (\code{\link[mapdeck]{add_pointcloud}})
}
\examples{
\dontrun{

cluster_deck(df,
  tooltip = "cluster_grp", fill_colour = "cluster_grp",
  elevation = NULL, radius = 10, palette = "viridis", legend = TRUE
)
}
}
