% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_aggregate.R
\name{dt_aggregate}
\alias{dt_aggregate}
\title{Aggregate data by time unit}
\usage{
dt_aggregate(
  df,
  dt_field = NULL,
  unit = "5 seconds",
  summary_fun = "median",
  na_rm = TRUE
)
}
\arguments{
\item{df}{a data frame with a datetime field.}

\item{dt_field}{character; name of datetime field.}

\item{unit}{character; string specifying a time unit or a multiple of a unit to be rounded}

\item{summary_fun}{character; summary function (i.e. 'mean', 'median'). Default = 'median.'}

\item{na_rm}{logical; should NA values be removed before the chosen
\code{summary_fun} computes?}
}
\description{
\code{dt_aggregate()} aggregates numeric values to a specified time unit. Columns with
uniform character values are retained (i.e. patient ID, sensor name).
}
\examples{
\dontrun{

dt_aggregate(df,
  dt_field = "Date_Time",
  unit = "5 seconds",
  summary_fun = "median", na_rm = TRUE
)
}
}
