% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_time.R
\name{filter_time}
\alias{filter_time}
\title{Filter time series data by a specific time interval}
\usage{
filter_time(
  df,
  t_var = NULL,
  t1 = NULL,
  t2 = NULL,
  span_days = FALSE,
  summary_tbl = TRUE
)
}
\arguments{
\item{df}{a data frame with a datetime  or time column.}

\item{t_var}{character; name of time or datetime column to filter.}

\item{t1}{character; start time of filtering interval. Start and and times should be in 'HH:MM:SS' format.}

\item{t2}{character; end time of filtering interval.}

\item{span_days}{logical; does the filtering interval include midnight? Default = FALSE.}

\item{summary_tbl}{logical; if TRUE (default), a summary table is included in the output showing the
number of observations occurring in each hour of the filtering interval.}
}
\description{
\code{filter_time()} can be used to filter time series data according to a specific time interval.
Traditionally, filtering by time can be messy due to the circular nature of time and
the 24 hour clock. \code{filter_time()} uses circular statistics to convert time to degrees, which enables
time series data to be filtered by time intervals that include midnight and span different
dates (i.e. '23:00:00' to '04:00:00') without issue. Because \code{filter_time()} is agnostic
to date, time series data can easily be filtered across multiple days.
}
\examples{

  filter_time(zoo_trip, t_var = 'Date_Time', t1 = '13:00:00', t2 = '14:30:00',
              summary_tbl = TRUE, span_days = TRUE)


}
