% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{friendship}
\alias{friendship}
\title{Walk through Friendship Park}
\format{
A data frame with 3,156 rows and 3 variables:
\describe{
\item{Date_Time}{datetime of logged coordinates}
\item{lat}{latitude}
\item{lon}{longitude}
}
}
\usage{
friendship
}
\description{
A dataset containing latitude and longitude coordinates recorded during a walk in
Theodore M. Berry International Friendship Park in Cincinnati, OH
}
\keyword{datasets}
