% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_places.R
\name{get_places}
\alias{get_places}
\title{Extract locations of stationary activity}
\usage{
get_places(
  df,
  cluster_col = "sp_temporal_cluster",
  nested = TRUE,
  geometry = TRUE,
  summary = FALSE,
  dt_field = NULL
)
}
\arguments{
\item{df}{a data frame with spatiotemporal or merged cluster groupings.}

\item{cluster_col}{character; name of column indicating cluster group (\code{sp_temporal_cluster} or \code{spatial_cluster})}

\item{nested}{logical; if TRUE, metadata for each cluster is nested in a list
column (\code{p_data})}

\item{geometry}{logical; if TRUE, an \code{sf} tibble with a corresponding \code{geometry}
column is returned}

\item{summary}{logical; if TRUE, the returned data frame is appended with summary
values for duration, start time, and end time for stationary activity recorded
for each cluster}

\item{dt_field}{character; name of datetime field used to calculate summary
information}
}
\value{
a data frame or \code{sf} object
}
\description{
\code{get_places()} analyzes periods of stationary activity within a cluster
and transforms the coordinates into a single location or 'place'. The input data frame may include
spatiotemporal cluster groupings generated by the \code{circleclust()} function or
spatially joined clusters (i.e. \code{merge_clusters()}). \code{get_places()} also summarizes the
duration of time (hours) spent as well as the first and last timestamp recorded at each location.
}
\examples{
\dontrun{

get_places(df, cluster_col = NULL, nested = TRUE, geometry = TRUE, summary = TRUE, dt_field = NULL)
}

}
