% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_clusters.R
\name{merge_clusters}
\alias{merge_clusters}
\title{Aggregate spatiotemporal clusters}
\usage{
merge_clusters(df, distance_threshold = 100, dt_field = NULL)
}
\arguments{
\item{df}{a data frame created by \code{circleclust()} with a \code{cluster_grp} column.
The data frame must also include datetime, longitude, and latitude columns.}

\item{distance_threshold}{numeric; distance threshold (meters) used to aggregate
clusters.}

\item{dt_field}{POSIXct; name of datetime field.}
}
\value{
a data frame. The original spatiotemporal cluster values are retained
in a column called \code{spatiotemp_cluster_grp}. Aggregated cluster values are
listed under \code{cluster_grp}.
}
\description{
\code{merge_clusters()} aggregates spatiotemporal clusters within a specific distance.
Essentially, clusters are aggregated spatially and the temporal combonent is removed.
}
\examples{
\dontrun{

merge_clusters(df,
  distance_threshold = 100, dt_field = NULL)
}
}
