% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bearings.R
\name{plot_bearings}
\alias{plot_bearings}
\title{Plot Bearings}
\usage{
plot_bearings(df, dt_field = NULL, divisor = NULL, export_options = FALSE)
}
\arguments{
\item{df}{a data frame created by \code{circleclust()} containing column \code{activity_status}}

\item{dt_field}{character; name of datetime field. If specified, the coordinate
timestamp is displayed in the plot tooltip. Default = NULL.}

\item{divisor}{numeric; If specified, the input data frame is reduced by a
a factor of 1/divisor. A divisor of 2 will omit every other row of the input
data frame reducing the data by half. This can be useful to visualize large
data sets. Default = NULL.}

\item{export_options}{logical; If TRUE, a drop down menu will be shown in plot
area with options to export a static image. Default = FALSE.}
}
\value{
a \code{highcharter} vector plot.
}
\description{
\code{plot_bearings} generates an interactive plot of personal sampling locations
using arrow markers. The direction of the arrows indicate the bearing between
sequential coordinates. Markers are characterized by activity status (mobile
vs. stationary).
}
\details{
See (\code{\link[highcharter]{hchart}})
}
\examples{
\dontrun{
plot_bearings(df, dt_field = NULL, divisor = NULL, export_options = FALSE )
}

}
