% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_token.R
\name{set_token}
\alias{set_token}
\title{Set \code{mapdeck} token}
\usage{
set_token(token)
}
\arguments{
\item{token}{mapbox access token}
}
\description{
Makes mapbox API token available for \code{mapdeck} calls
}
\details{
See (\code{\link[mapdeck]{set_token}}).

You can request a mapbox acces token \href{https://docs.mapbox.com/help/how-mapbox-works/access-tokens/}{here}.
}
\examples{
\dontrun{

set_token("my_mapbox_token")
}
}
