% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgs_sf.R
\name{wgs_sf}
\alias{wgs_sf}
\title{Create \code{sf} Object}
\usage{
wgs_sf(df)
}
\arguments{
\item{df}{a data frame with columns for latitude (\code{lat}) and longitude (\code{lon}).}
}
\value{
a \code{POINT} object.
}
\description{
Creates a multipoint \code{sf} object. Points are projected using WGS84 (crs = 4326).
Missing coordinates are allowed, but given an empty geometry.
}
\examples{
\dontrun{

ufp_sf(df)
}
}
