% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{woodland}
\alias{woodland}
\title{Car transit and walking trip through Woodland Mound}
\format{
A data frame with 1,279 rows and 3 variables:
\describe{
\item{Date_Time}{datetime of logged coordinates}
\item{lat}{latitude}
\item{lon}{longitude}
}
}
\usage{
woodland
}
\description{
A dataset containing latitude and longitude coordinates recorded during car
transit and a walk through Woodland Mound in Cincinnati, OH
}
\keyword{datasets}
