% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zoo_trip}
\alias{zoo_trip}
\title{Walk through Cincinnati Zoo and Botanical Gardens}
\format{
A data frame with 10,363 rows and 3 variables:
\describe{
\item{Date_Time}{datetime of logged coordinates}
\item{lat}{latitude}
\item{lon}{longitude}
}
}
\usage{
zoo_trip
}
\description{
A dataset containing latitude and longitude coordinates recorded during a walk through
the Cincinnati Zoo and Botanical Gardens.
}
\keyword{datasets}
