% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circleclust.R
\name{circleclust}
\alias{circleclust}
\title{Cluster coordinates based on circular variance}
\usage{
circleclust(
  df,
  dt_field = NULL,
  circvar_threshold = 0.7,
  window = 60,
  show_circvar = FALSE,
  rspeed_threshold = NULL,
  pl_dist_threshold = NULL,
  cluster_threshold = NULL
)
}
\arguments{
\item{df}{a data frame with columns \code{speed} and \code{azimuth} created by \code{move()}.
The data frame must also include datetime, longitude, and latitude colums.}

\item{circvar_threshold}{numeric; circular variance threshold to determine clustering.
Default = 0.7.}

\item{window}{numeric; window (number of rows) in which to calculate circular variance.}

\item{show_circvar}{logical; if \code{TRUE}, a column will be added to the output data
frame listing the circular variance (\code{circvar}) for each observation.
Default = \code{FALSE}.}

\item{rspeed_threshold}{numeric; if assigned a numeric value, the 1-minute rolling
median speed (m/s) is calculated. Observations with a circular variance above
'circvar_threshold' and below 'rspeed_threshold' are assigned to a cluster. Assigning
a value to this parameter can mitigate incidental clustering of coordinates due
to stop-and-go transit.}

\item{pl_dist_threshold}{numeric; distance threshold (meters) used to aggregate
clusters. If the distance between consecutive clusters is below this threshold,
each cluster, and the coordinates between them, are combined into a single
grouping of coordinates. Setting this parameter may be useful if location data
includes transit at a low speed with frequent stops (i.e. walking).}

\item{cluster_threshold}{numeric; the minimum allowable number of observations
in each cluster.  If the number of observations in an identified cluster is less than this threshold,
the observations are retained but not assigned a to a cluster.}
}
\value{
a tibble.
}
\description{
\code{circleclust} calculates the circular variance of consecutive latitude and
longitude coordinates within a moving window and identifies spatiotemporal clusters
based on a set of flexible parameters.  These parameters include the size of the
window in which to calculate circular variance, the threshold value (circular variance)
to identify clustering, and the minimum number of observations allowed in a
cluster.
}
\details{
Observations belonging to a cluster are assigned to a cluster group (\code{cluster_grp}),
which are ordered temporally. Further, observations are marked as either
static or mobile (\code{activity_status}).

Imputing lon/lat values using \code{impute_coords()} is recommended if GPS
coordinates are missing.
}
\examples{
\dontrun{

circleclust(df, circvar_threshold = .7, window = 60, cluster_threshold = NULL,
  show_circvar = FALSE, rspeed_threshold = NULL,
  pl_dist_threshold = NULL,  cluster_threshold = NULL)
}
}
