% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_aggregate.R
\name{dt_aggregate}
\alias{dt_aggregate}
\title{Aggregate PUFP data}
\usage{
dt_aggregate(
  df,
  dt_field = NULL,
  unit = "5 seconds",
  floor_or_celiling = "floor",
  summary_fun = "median"
)
}
\arguments{
\item{df}{a data frame with a datetime field.}

\item{dt_field}{character; name of datetime field.}

\item{unit}{character; string specifying a time unit or a multiple of a unit to be rounded}

\item{floor_or_celiling}{character; either 'floor'(\code{\link[lubridate]{floor_date}})
or 'ceiling'(\code{\link[lubridate]{ceiling_date}}). Default = 'floor.'}

\item{summary_fun}{character; summary function (i.e. 'mean', 'median'). Default = 'median.'}
}
\value{
a tibble
}
\description{
\code{dt_aggregate()} aggregates numeric values to a specified time unit. Columns with
unique character values are retained (i.e. patient ID, sensor name).
}
\examples{
\dontrun{

dt_aggregate(df, dt_field = 'Date_Time',
  unit = "5 seconds", floor_or_ceiling = "floor",
  summary_fun = "median"
)
}
}
