% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_places.R
\name{get_places}
\alias{get_places}
\title{Extract locations of stationary activity}
\usage{
get_places(
  df,
  nested = TRUE,
  geometry = TRUE,
  summary = FALSE,
  dt_field = NULL
)
}
\arguments{
\item{df}{a data frame created by \code{circleclust()} containing cluster group (\code{cluster_grp})}

\item{nested}{logical; if TRUE, metadata for each cluster is nested in a list
column (\code{p_data})}

\item{geometry}{logical; if TRUE, an \code{sf} tibble with a corresponding \code{geometry}
column is returned}

\item{summary}{logical; if TRUE, the returned data frame is appended with summary
values for duration, start time, and end time for stationary activity recorded
for each cluster}

\item{dt_field}{character; name of datetime field used to calculate summary
information}
}
\value{
a data frame or \code{sf} object
}
\description{
\code{get_places()} analyzes periods of stationary activity within each spatiotemporal
cluster identified by \code{circleclust()} and transforms the coordinates into a
single location or 'place'.
}
\examples{
\dontrun{

get_places(df, nested = TRUE, geometry = TRUE, summary = TRUE, dt_field = NULL)
}

}
