% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tracks.R
\name{get_tracks}
\alias{get_tracks}
\title{Extract periods of mobile activity}
\usage{
get_tracks(df, nested = TRUE, geometry = TRUE, summary = TRUE, dt_field = NULL)
}
\arguments{
\item{df}{a data frame created by \code{circleclust()} containing cluster group (\code{cluster_grp})}

\item{nested}{logical; if TRUE, metadata for each track is nested in a list
column (\code{t_data})}

\item{geometry}{logical; if TRUE, an \code{sf} tibble with a corresponding \code{geometry}
column is returned. Tracks are converted into a \code{LINESTRING}, which connect the
beginning and ending coordinate of each track.}

\item{summary}{logical; if TRUE, the returned data frame is appended with summary
values for duration, start time, and end time for each track. The
dominant bearing of each track is also calculated.}

\item{dt_field}{character; name of datetime field used to calculate summary
information.}
}
\value{
a data frame or \code{sf} object
}
\description{
\code{get_tracks()} extracts periods of mobile activity or tracks identified by
\code{circleclust()}
}
\examples{
 \dontrun{
get_tracks(df, nested = TRUE, geometry = TRUE, summary = TRUE, dt_field = NULL)
}

}
