% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_time.R
\name{impute_time}
\alias{impute_time}
\title{Impute datetime for gaps in GPS data}
\usage{
impute_time(df, dt_field = NULL)
}
\arguments{
\item{df}{a data frame with a datetime field.}

\item{dt_field}{character; name of datetime field.}
}
\value{
data frame.
}
\description{
\code{impute_time()} imputes datetime for gaps in GPS data.  Gaps in time may result
from filtering of invalid GPS data or because the GPS logger goes into 'sleep'
mode (i.e. Columbus GPS logger). Datetime rows are added to the input data frame
under the specified datetime column (\code{dt_field}) to fill gaps between time stamps.
The sampling interval is automatically calculated.
}
\details{
\code{impute_coords()} can be used in conjunction with \code{impute_time()} to assign
latitude and longitude to the added time stamps/rows.
}
\examples{
\dontrun{
impute_time(df, dt_field = 'Date_Time') \%>\%
  impute_coords()
}

}
